Crate hotg_runecoral

Bindings for exposing the functionality of librunecoral

Example

use hotg_runecoral::{Tensor, TensorMut, InferenceContext, AccelerationBackend};

// load the model
let model: &[u8] = load_model();

// set aside some arrays for our inputs and outputs
let input = [0.0_f32];
let mut output = [0.0_f32];

// And create tensors which point to them
let input_tensor = Tensor::from_slice(&input, &[1]);
let output_tensor = TensorMut::from_slice(&mut output, &[1]);

// load our inference backend
let mut ctx = InferenceContext::create_context(
    "application/tflite-context",
    model,
    AccelerationBackend::NONE,
)?;

// Now we can run inference
ctx.infer(&[input_tensor], &mut [output_tensor])?;

// and look at the results
println!("{:?} => {:?}", input, output);

Modules

ffi

The raw bindings generated by bindgen.

Structs

AccelerationBackend

InferenceContext

A backend which can run inference on a model.

Tensor

An immutable reference to a tensor’s backing buffer.

TensorDescriptor

The shape and element type of a Tensor.

TensorMut

A mutable reference to a tensor’s backing buffer.

Enums

ElementType

Possible element types that can be used in a Tensor.

Error

LoadError

Traits

TensorElement

A Rust type that can be used as the element of a tensor.

Functions

mimetype

The mimetype used by this crate to represent TensorFlow Lite models.